requirements.txt
# Core dependencies for NeRF MLP training and rendering
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
matplotlib>=3.5.0
scikit-image>=0.19.0
psutil>=5.8.0
tqdm>=4.64.0

# Optional dependencies for specific features
imageio>=2.19.0  # For image I/O in some scripts
tensorflow>=2.8.0  # Only needed for example/ directory (original NeRF)

# Development/testing (optional)
jupyter>=1.0.0  # For notebooks if added later
